Editor: Update selected text logic for key up
This updates the logic used on keyup in the editor to grab the selected
text. We need to fire the getSelectedText function on any key up. This
ensures that the correct word count is shown when the words are
highlighted with a keyboard combination and also clears out the word
count if necessary on key up. This shouldn't be an issue because we
check whether or not state should be updated in getSelectedText.